requirements.txt
scipy
matplotlib
numpy
pandas
Pillow